- Add ERD diagram - Refactor URLs to has more sense - Split in 2 controllers - Update postman collection
