+++ src/main/java/com/emmthias/cache/controller/CacheBucketObjectController.java
@@ -1,12 +1,10 @@
 package com.emmthias.cache.controller;
 
-
 import com.emmthias.cache.common.response.error.IApiErrorResponse;
 import com.emmthias.cache.common.response.success.IApiResponse;
 import com.emmthias.cache.common.response.success.impl.GetCacheResponse;
 import com.emmthias.cache.domain.Cache;
 import com.emmthias.cache.domain.CacheObject;
-import com.emmthias.cache.domain.CachePreference;
 import com.emmthias.cache.service.ICacheService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -14,14 +12,12 @@ import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import org.json.JSONException;
-import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -34,38 +30,29 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
-import static com.emmthias.cache.constants.Constants.CACHE_BUCKET_URL;
-import static com.emmthias.cache.constants.Constants.CACHE_SERVICES_ENDPOINT;
-import static com.emmthias.cache.constants.Constants.CACHE_URL;
+import static com.emmthias.cache.constants.Constants.CACHE_SERVICE_BUCKET_ENDPOINT;
+import static com.emmthias.cache.constants.Constants.CACHE_OBJECT_URL;
 import static com.emmthias.cache.constants.Constants.KEY;
-import static com.emmthias.cache.constants.Constants.OBJECT_URL;
-import static com.emmthias.cache.constants.Constants.REQUEST_TYPE_DELETE;
 import static com.emmthias.cache.constants.Constants.REQUEST_TYPE_GET;
 import static com.emmthias.cache.constants.Constants.REQUEST_TYPE_PATCH;
 import static com.emmthias.cache.constants.Constants.REQUEST_TYPE_POST;
 import static com.emmthias.cache.constants.Constants.REQUEST_TYPE_PUT;
-import static com.emmthias.cache.converter.Converters.ConvertToJsonElement;
 import static com.emmthias.cache.converter.Converters.buildGetCacheResponse;
-import static com.emmthias.cache.converter.Converters.convertMapIntoJsonStructure;
 import static com.emmthias.cache.converter.Converters.convertMapToCacheObject;
 import static com.emmthias.cache.converter.Converters.convertMapToSingleCacheObject;
-import static com.emmthias.cache.validator.ValidatorUtil.validateCachePreference;
 
 /**
- * Controller to handle cache service
+ * Controller to handle cache object service
  * <p>
- * End point that handle the cache service operations.
+ * End point that handle the cache object service operations.
  * </p>
  */
 
-// TODO
-// Add unit testing
-// split into small projects
 @RestController
-@RequestMapping({CACHE_SERVICES_ENDPOINT})
+@RequestMapping({CACHE_SERVICE_BUCKET_ENDPOINT})
 @Api("cache service operation controller")
-public class CacheController {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CacheController.class);
+public class CacheBucketObjectController {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CacheBucketObjectController.class);
 
     @Autowired
     ICacheService cacheService;
@@ -82,35 +69,7 @@ public class CacheController {
                     response = IApiErrorResponse.class)
     })
     @PostMapping(name = "Create an cache object by key",
-            value = {CACHE_BUCKET_URL},
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<IApiResponse> createBucketByKey(
-            @ApiParam(required = true, name = "key", value = "the key identifier")
-            @PathVariable(KEY) String key,
-            @RequestBody @ApiParam(required = true,
-                    allowEmptyValue = false, name = "data", value = "The request body content ")
-                    Cache element) throws JSONException, ExecutionException, InterruptedException {
-        CachePreference preference = validateCachePreference(element.getCachePreference());
-        List<CacheObject> data = convertMapToCacheObject(element.getElements());
-
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(buildGetCacheResponse(cacheService.createCacheObject(key, data, preference).get()));
-    }
-
-    @ApiOperation(value = "This end point will insert the {object} provided in the body of the request into a slot in" +
-            "memory at {key}",
-            httpMethod = REQUEST_TYPE_POST)
-    @ApiResponses(value = {
-            @ApiResponse(code = 201,
-                    message = "The created object added into a cache layer.",
-                    response = GetCacheResponse.class),
-            @ApiResponse(code = 507,
-                    message = "The cache layer has Insufficient Storage.",
-                    response = IApiErrorResponse.class)
-    })
-    @PostMapping(name = "Create an cache object by key",
-            value = {CACHE_URL},
+            value = {CACHE_OBJECT_URL},
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<IApiResponse> addBucketObjectByKey(
@@ -135,7 +94,7 @@ public class CacheController {
                     message = "Not found element",
                     response = IApiErrorResponse.class)})
     @GetMapping(name = "Retrieve cache object by key",
-            value = {CACHE_URL},
+            value = {CACHE_OBJECT_URL},
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<IApiResponse> getBucketByKey(
@@ -146,27 +105,6 @@ public class CacheController {
                 .body(buildGetCacheResponse(cacheService.getCacheObjectByKey(key).get()));
     }
 
-    @ApiOperation(value = "This end point will return all the cache objects.",
-            httpMethod = REQUEST_TYPE_GET)
-    @ApiResponses(value = {
-            @ApiResponse(code = 200,
-                    message = "The elements belongs to a cache layer identifying by key.",
-                    response = GetCacheResponse.class),
-            @ApiResponse(code = 404,
-                    message = "Not found element",
-                    response = IApiErrorResponse.class)})
-    @GetMapping(name = "Retrieve cache object by key",
-            value = {OBJECT_URL},
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<IApiResponse> getAllBuckets() throws JSONException, ExecutionException, InterruptedException {
-
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(buildGetCacheResponse(
-                        convertMapIntoJsonStructure(cacheService.getMapInstance().get()))
-                );
-    }
-
     @ApiOperation(value = "This end point will patch the {object} provided in the body making the merge of " +
             "store information with new value in the {object} of the request into their corresponding slot in memory " +
             "at {key}",
@@ -175,7 +113,7 @@ public class CacheController {
             @ApiResponse(code = 200,
                     message = "The element associated to cache layer identifying by key",
                     response = GetCacheResponse.class)})
-    @PatchMapping(name = "patch", value = {CACHE_URL}, consumes = MediaType.APPLICATION_JSON_VALUE,
+    @PatchMapping(name = "patch", value = {CACHE_OBJECT_URL}, consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<IApiResponse> patchBucketObjectByKey(
             @ApiParam(required = true, name = "key", value = "the key identifier")
@@ -195,7 +133,7 @@ public class CacheController {
             @ApiResponse(code = 200,
                     message = "The element associated to cache layer identifying by key",
                     response = GetCacheResponse.class)})
-    @PutMapping(name = "put", value = {CACHE_URL}, consumes = MediaType.APPLICATION_JSON_VALUE,
+    @PutMapping(name = "put", value = {CACHE_OBJECT_URL}, consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<IApiResponse> updateBucketObjectByKey(
             @ApiParam(required = true, name = "key", value = "the key identifier")
@@ -206,26 +144,5 @@ public class CacheController {
         CacheObject elementToUpdate = convertMapToSingleCacheObject(element.getElement());
         return ResponseEntity.status(HttpStatus.OK)
                 .body(buildGetCacheResponse(cacheService.updateObject(key, elementToUpdate).get()));
-    }
-
-    @ApiOperation(value = "This end point will delete the object stored at slot {key}.",
-            httpMethod = REQUEST_TYPE_DELETE)
-    @ApiResponses(value = {
-            @ApiResponse(code = 200,
-                    message = "The deleted element to a cache layer identifying by key",
-                    response = GetCacheResponse.class)})
-    @DeleteMapping(name = "This end point will delete the object stored at slot {key}", value = {CACHE_URL},
-            consumes = MediaType.APPLICATION_JSON_VALUE,
-            produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<IApiResponse> deleteBucketByKey(
-            @ApiParam(required = true, name = "key", value = "the key identifier")
-            @PathVariable(KEY) String key,
-            @RequestBody @ApiParam(required = true,
-                    allowEmptyValue = false, name = "codeDetailRequest", value = "The request body content ")
-                    JSONObject element) throws JSONException, ExecutionException, InterruptedException {
-        return ResponseEntity.status(HttpStatus.OK).body(
-                buildGetCacheResponse(
-                        ConvertToJsonElement("response", cacheService.delete(key).get()))
-        );
     }
 }
